requirements.txt
Flask==2.1.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==0.21.1
Flask-Mail==0.9.1
APScheduler==3.9.1
sendgrid==6.11.0
Werkzeug==2.0.3